update utils  & add unit test v2 models get_latest_model_info
